separate handleInactiveClose
