+++ src/onInterval.js
@@ -1,4 +1,5 @@
 // @ts-check
+import { handleInactiveClose } from './handleInactiveClose.js'
 /**
  * @typedef {import('./logger.js').Logger} Logger
  * @typedef {import('discord.js').AnyThreadChannel} AnyThreadChannel
@@ -16,7 +17,7 @@ export async function onInterval(logger, forums) {
 
   await Promise.all(
     forums.map(forum =>
-      archiveInactiveThreads(logger.createChild('WatchInactiveThread'), forum)
+      onIntervalForForum(logger.createChild(`Forum:${forum.setting.id}`), forum)
     )
   )
 
@@ -27,7 +28,7 @@ export async function onInterval(logger, forums) {
  * @param {Logger} logger
  * @param {Forum} param1
  */
-async function archiveInactiveThreads(logger, { channel, setting }) {
+async function onIntervalForForum(logger, { channel, setting }) {
   const { threads: activeThreads } = await channel.threads.fetchActive()
 
   logger.info(`Found ${activeThreads.size} active threads`)
@@ -35,7 +36,7 @@ async function archiveInactiveThreads(logger, { channel, setting }) {
   const inactiveDurationDay = 2
   const results = await Promise.all(
     activeThreads.map(thread =>
-      archiveIfInactive(thread, inactiveDurationDay, setting)
+      handleInactiveClose(thread, inactiveDurationDay, setting)
     )
   )
   const archived = results.filter(archived => archived).length
@@ -43,23 +44,4 @@ async function archiveInactiveThreads(logger, { channel, setting }) {
   logger.info(
     `Found ${archived} threads over ${inactiveDurationDay} day since last activity.`
   )
-}
-
-/**
- * @param {AnyThreadChannel} thread
- * @param {number} inactiveDurationDay
- * @param {ForumChannelSetting} setting
- * @returns {Promise<boolean>} true if archived
- */
-async function archiveIfInactive(thread, inactiveDurationDay, setting) {
-  const messages = await thread.messages.fetch({ limit: 1 })
-  const lastMessage = messages.first()
-  if (!lastMessage) return false
-
-  const inactiveDuration = inactiveDurationDay * (1000 * 60 * 60 * 24)
-  if (Date.now() - lastMessage.createdTimestamp < inactiveDuration) return false
-
-  await thread.send(setting.onStale(thread.ownerId, inactiveDurationDay))
-  await thread.setArchived(true, `${inactiveDurationDay}日間操作がなかったため`)
-  return true
 }
